Fix local variables not being marked as such.
A problem was in the fact that LuaJIT marks local variable
starting with a instruction after a LOGICAL assignment block. I.e. after
all the instructions emitted by the assignment block.

That is: local a = 1 will setup the debuginfo for the "a" variable
starting from the last-emited + 1 instruction, but
local a1, a2, ..., a100 = 1, 2, ..., 100 will setup the debuginfo for
all the a* variables starting from the last-emited + 100 instruction.

The solution is to defer the locality assumption until the first slot
use. This way we can't mark an assignment node as being a local
definition so we have to move this functionality into the optimizer