Replace UCLO's pointing to returns with the returns
It makes sense if you'll think about it - if you are trying to return
above of a function declaration, you still need to close stuff, so
you will be using UCLO here. And where to jump then - it could be
a very special return instruction here. Right - let's add a new return
instruction at the end of the function and jump there.

Who cares about a broken flow, really...