builder: a bit of refactoring for a warps build process
Now warps are built in their own fancy loop, before any other instruction
is processed. This commit was inteded to fix some bugs with
if bla then bla else end

type statements (with an empty but preset else clause), but it turned out
that it won't help, so this change isn't fixing anything. But it's a bit
more "correct", so I'll commit it anyway